fix: Résolution build Next.js 15 avec wrapper Client Component
• PublicationViewer.tsx: Nouveau wrapper Client Component avec 'use client'
• page.tsx: Migration import dynamique vers wrapper local
• PdfClientViewer.tsx: Correction worker PDF.js avec unpkg CDN
• Architecture: Préservation Server Component + isolation logique client

🔧 Résout erreur "ssr: false not allowed in Server Components"
✅ Build Vercel: ✓ Compiled successfully in 23.0s
