+++ public/graph-billets.json
@@ -277,7 +277,7 @@
     }
   ],
   "metadata": {
-    "generatedAt": "2025-08-16T14:21:12.915Z",
+    "generatedAt": "2025-08-16T15:02:37.334Z",
     "billetsCount": 26,
     "nodesCount": 26,
     "stability": {
